requirements.txt
# Core Framework
streamlit==1.31.0

# LLM Integration
langchain-groq==0.1.0

# Embeddings and ML
sentence-transformers==2.3.1
transformers==4.36.0

# Vector Database
faiss-cpu==1.7.4
numpy==1.24.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Environment Management
python-dotenv==1.0.0

# Additional Dependencies
pillow==10.2.0
typing-extensions==4.9.0
